requirement.txt
mindspore-gpu==1.5.0rc1
numpy
argparse
